+++ rocketgram/api/get_chat_member.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 from typing import Union
 
 from .request import Request
-from .. import api
+from .. import api, context
 
 
 @dataclass(frozen=True)
@@ -24,6 +24,6 @@ class GetChatMember(Request):
         assert isinstance(data, dict), "Should be dict."
         return api.ChatMember.parse(data)
 
-    async def send2(self) -> 'api.ChatMember':
-        res = await self._send()
+    async def send(self) -> 'api.ChatMember':
+        res = await context.bot.send(self)
         return res.result