+++ rocketgram/api/get_my_commands.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 from typing import List
 
 from .request import Request
-from .. import api
+from .. import api, context
 
 
 @dataclass(frozen=True)
@@ -21,6 +21,6 @@ class GetMyCommands(Request):
         assert isinstance(data, list), "Should be list."
         return [api.BotCommand.parse(r) for r in data]
 
-    async def send2(self) -> List['api.BotCommand']:
-        res = await self._send()
+    async def send(self) -> List['api.BotCommand']:
+        res = await context.bot.send(self)
         return res.result