+++ rocketgram/api/get_sticker_set.py
@@ -6,7 +6,7 @@
 from dataclasses import dataclass
 
 from .request import Request
-from .. import api
+from .. import api, context
 
 
 @dataclass(frozen=True)
@@ -22,6 +22,6 @@ class GetStickerSet(Request):
         assert isinstance(data, dict), "Should be dict."
         return api.StickerSet.parse(data)
 
-    async def send2(self) -> 'api.StickerSet':
-        res = await self._send()
+    async def send(self) -> 'api.StickerSet':
+        res = await context.bot.send(self)
         return res.result