+++ rocketgram/routers/dispatcher/__init__.py
@@ -8,5 +8,4 @@ from . import commonwaiters
 from .base import BaseDispatcher
 from .dispatcher import Dispatcher
 from .filters import make_filter, priority
-from .proxy import BaseDispatcherProxy
 from .waiters import make_waiter